Separate games into multiple files
